move token-related utilities and types to own file
