+++ packages/web-client/app/components/card-pay/deposit-workflow/transaction-setup/index.ts
@@ -11,8 +11,8 @@ import { toBN } from 'web3-utils';
 import {
   TokenDisplayInfo,
   TokenSymbol,
-} from '@cardstack/web-client/utils/web3-strategies/token-display-info';
-import { bridgeableSymbols } from '@cardstack/web-client/utils/web3-strategies/token-categories';
+  bridgeableSymbols,
+} from '@cardstack/web-client/utils/token';
 
 interface CardPayDepositWorkflowTransactionSetupComponentArgs {
   workflowSession: WorkflowSession;
