+++ packages/web-client/app/services/layer2-network.ts
@@ -8,8 +8,7 @@ import { reads } from 'macro-decorators';
 import WalletInfo from '../utils/wallet-info';
 import BN from 'bn.js';
 import { DepotSafe } from '@cardstack/cardpay-sdk/sdk/safes';
-import { ConversionFunction } from '@cardstack/web-client/utils/web3-strategies/types';
-import { ConvertibleSymbol } from '@cardstack/web-client/utils/web3-strategies/token-categories';
+import { ConvertibleSymbol, ConversionFunction } from '@cardstack/web-client/utils/token';
 
 export default class Layer2Network extends Service {
   strategy!: Layer2Web3Strategy;
