+++ packages/web-client/app/utils/token.ts
@@ -1,5 +1,41 @@
-export type TokenSymbol = 'DAI' | 'CARD' | 'ETH';
+import { AbiItem } from 'web3-utils';
+import { ERC20ABI } from '@cardstack/cardpay-sdk/index.js';
+import { getAddressByNetwork } from '@cardstack/cardpay-sdk';
+import { ChainAddress } from './web3-strategies/types';
 
+// symbols
+export type ConvertibleSymbol = 'DAI' | 'CARD';
+export type BridgeableSymbol = 'DAI' | 'CARD';
+export type Layer1BalanceSymbol = 'DAI' | 'CARD' | 'ETH';
+export type TokenSymbol =
+  | ConvertibleSymbol
+  | BridgeableSymbol
+  | Layer1BalanceSymbol;
+
+// conversion
+// eslint-disable-next-line no-unused-vars
+export type ConversionFunction = (amountInWei: string) => number;
+export const convertibleSymbols: ConvertibleSymbol[] = ['CARD', 'DAI'];
+
+// contract/bridging
+export const bridgeableSymbols: BridgeableSymbol[] = ['CARD', 'DAI'];
+const contractNames: Record<BridgeableSymbol, string> = {
+  DAI: 'daiToken',
+  CARD: 'cardToken',
+};
+
+export class TokenContractInfo {
+  symbol: BridgeableSymbol;
+  address: ChainAddress;
+  abi = ERC20ABI as AbiItem[];
+
+  constructor(symbol: BridgeableSymbol, network: string) {
+    this.symbol = symbol;
+    this.address = getAddressByNetwork(contractNames[this.symbol], network);
+  }
+}
+
+// display only
 interface DisplayInfo {
   name?: string;
   symbol: TokenSymbol;
