+++ packages/web-client/app/utils/web3-strategies/test-layer2.ts
@@ -1,12 +1,10 @@
 import { tracked } from '@glimmer/tracking';
 import WalletInfo from '../wallet-info';
+import { ChainAddress, Layer2Web3Strategy, TransactionHash } from './types';
 import {
-  ChainAddress,
+  ConvertibleSymbol,
   ConversionFunction,
-  Layer2Web3Strategy,
-  TransactionHash,
-} from './types';
-import { ConvertibleSymbol } from '@cardstack/web-client/utils/web3-strategies/token-categories';
+} from '@cardstack/web-client/utils/token';
 import RSVP, { defer } from 'rsvp';
 import BN from 'bn.js';
 import { fromWei, toBN } from 'web3-utils';
