--- packages/web-client/app/utils/web3-strategies/token-categories.ts
@@ -1,6 +1,0 @@
-export const bridgeableSymbols: TokenSymbol[] = ['CARD', 'DAI'];
-
-export type ConvertibleSymbol = 'DAI' | 'CARD';
-export const convertibleSymbols: ConvertibleSymbol[] = ['CARD', 'DAI'];
-
-export type TokenSymbol = 'DAI' | 'CARD' | 'ETH';
